requirements.txt
Click==7.0
Django==2.2.1
django-crispy-forms==1.7.2
mysqlclient==1.4.2.post1
objects==0.3.1
Pillow==6.0.
sqlparse==0.3.0
